tweak: normalize judge scores to 0.0-1.0 float system-wide
Replaces the manufactured 1-10 integer score scale (JudgeResult,
JudgeEvaluation, JudgedRecord) with gavel's intended 0.0-1.0 default.
DeepEval and external/custom judge scores now pass through natively
instead of being remapped; Autotune's normalize_score() and
TuneStep._compute_avg_score() no longer divide by 10 (that would have
silently double-normalized once the underlying scale changed). Report
templates and log lines updated to display the new scale.

+++ src/gavel_ai/reporters/jinja_reporter.py
@@ -171,7 +171,7 @@ class Jinja2Reporter(Reporter):
             return []
 
         # Aggregate scores by variant
-        variant_scores: Dict[str, List[int]] = {}
+        variant_scores: Dict[str, List[float]] = {}
 
         for result in results:
             variant_id = result.get("variant_id", "unknown")
